% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_jsons.R
\name{add_name_column}
\alias{add_name_column}
\title{Add 'name' column to unnested tables}
\usage{
add_name_column(x)
}
\arguments{
\item{x}{data.frame/tibble with \code{level_x} columns from unnesting}
}
\value{
x with additional column \code{name}, or if nesting is only one level deep
(i.e., \code{level_1}), the column is renamed
@noRd
}
\description{
Add 'name' column to unnested tables
}
