% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{clean_categories}
\alias{clean_categories}
\title{Identify categories from CRF trend tables}
\usage{
clean_categories(x)
}
\arguments{
\item{x}{character, "raw" categories read directly from CRF trend tables}
}
\value{
character, short-format categories (e.g., 1.A.2, 2.)
}
\description{
Identify categories from CRF trend tables
}
