% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{clean_categories_df}
\alias{clean_categories_df}
\title{Tabled CRF-Categories}
\usage{
clean_categories_df(x)
}
\arguments{
\item{x}{character, "raw" categories from CRF table}
}
\value{
data.frame with 5 columns
}
\description{
Tabled CRF-Categories
}
