% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{find_id}
\alias{find_id}
\title{Search through UNFCCC Ids}
\usage{
find_id(rms, id, verbose = FALSE)
}
\arguments{
\item{rms}{object from \code{rem_init()}}

\item{id}{integer}
}
\value{
character, name corresponding to id
}
\description{
Search through UNFCCC Ids
}
