% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_jsons.R
\name{flatten_dims}
\alias{flatten_dims}
\title{Recursive unnest of instance lists}
\usage{
flatten_dims(x, depth = 1)
}
\arguments{
\item{x}{list, a category list from an instance response, such as
"api/dimension-instances/classification".}

\item{depth}{numeric, depth of nestedness.}
}
\value{
tibble with columns \code{id}, \code{level_x}, with x reflecting the nestedness,
and \code{name}. \code{id} is the DI API id, values in \code{level_x} are the corresponding name
for the id, or its parent, and name is the id's name.
}
\description{
Instance responses return lists after parsing. Each item corresponds to an
UNFCCC / Worldbank category (e.g., annex-I). This function parses the information,
i.e., \code{name} and \code{id} for each element and sub-element (e.g., LULUCF - Deforestation)
into a tibble reflecting the nestedness.
}
