% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_jsons.R
\name{flatten_parties}
\alias{flatten_parties}
\title{Flatten party JSON-list response}
\usage{
flatten_parties(parties)
}
\arguments{
\item{parties}{}
}
\value{
data.frame of parties
}
\description{
Flatten party JSON-list response
}
