% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{flex_query}
\alias{flex_query}
\title{Query UNFCCC DI with Flex Query API}
\usage{
flex_query(
  rms,
  variable_ids,
  party_ids,
  year_ids,
  path = "api/records/flexible-queries/",
  pretty = TRUE
)
}
\arguments{
\item{rms}{list, object form \code{rem_init()}}

\item{variable_ids}{integer, ids of interest}

\item{party_ids}{integer, ids of interest}

\item{year_ids}{integer, ids of interest}

\item{path}{character, default is 'api/records/flexible-queries/'}

\item{pretty}{logical, should raw results transformed to feature text
descriptions instead of ids?}
}
\value{
data.frame
}
\description{
Query UNFCCC DI with Flex Query API
}
