% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{get_crf_level}
\alias{get_crf_level}
\title{Identify CRF-Category levels}
\usage{
get_crf_level(x)
}
\arguments{
\item{x}{character, CRF-categories in short-form (e.g., 1.A.2, or 2.) from \link(\code{clean_categories()})}
}
\value{
ch
}
\description{
Identify CRF-Category levels
}
