% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_names}
\alias{get_names}
\title{Get name based on id}
\usage{
get_names(rms, ids)
}
\arguments{
\item{rms}{list, lower-level item from \code{remis} object, like \code{..$parties}}

\item{ids}{}
}
\value{
character, names corresponding to \code{ids}
}
\description{
Get name based on id
}
