% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_parse}
\alias{get_parse}
\title{GET and parse objects}
\usage{
get_parse(rms, url)
}
\arguments{
\item{rms}{response from}

\item{url}{character, api path for desired object}
}
\value{
list, parsed from json
}
\description{
GET and parse objects
}
