% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_jsons.R
\name{make_idtbl}
\alias{make_idtbl}
\title{content placeholder}
\usage{
make_idtbl(nrow, depth)
}
\arguments{
\item{nrow}{numeric, number of rows for the placeholder \code{tibble}.
Should be the same size as all non-children items in extraction list.}

\item{depth}{numeric, corresponds to depth in nested extraction list.}
}
\value{
tibble with columns \code{id}, \code{level_x}, with x reflecting the nestedness,
and \code{name}. \code{id} is the DI API id, values in \code{level_x} are the corresponding name
for the id, or its parent, and name is the id's name.
}
\description{
Generate placeholder tibbles for
}
