% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{parse_raw}
\alias{parse_raw}
\title{Readable query results}
\usage{
parse_raw(rms, raw)
}
\arguments{
\item{rms}{list, \code{remis} object}

\item{raw}{list, response from POST request}
}
\value{
data.frame
}
\description{
This function combines numeric/character responses from queries with their
respective id values from parties, years and variables.
}
