% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{post_parse}
\alias{post_parse}
\title{POST and parse objects}
\usage{
post_parse(rms, path = "api/records/flexible-queries/", body, pretty = TRUE)
}
\arguments{
\item{rms}{list, remis object}

\item{path}{character}

\item{body}{list containing parties, years, variables}

\item{parse}{logical, should raw results transformed to feature text
descriptions instead of ids?}
}
\value{
data.frame, parsed from json
}
\description{
POST and parse objects
}
