% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{read_emission_trends}
\alias{read_emission_trends}
\title{Extract trend data from CRF Table}
\usage{
read_emission_trends(path, table = "Table10s1")
}
\arguments{
\item{path}{character, single file path to CRF table.}

\item{table}{character, table (i.e., sheet) name}
}
\value{
data.frame of CRF values
}
\description{
Extract trend data from CRF Table
}
