% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{req_parse}
\alias{req_parse}
\title{Get and parse objects}
\usage{
req_parse(x, url)
}
\arguments{
\item{x}{response from}

\item{url}{character, api path for desired object}
}
\value{
list, parsed from json
}
\description{
Get and parse objects
}
