% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_processing.R
\name{set_data_numeric}
\alias{set_data_numeric}
\title{Creates numeric data from raw tables}
\usage{
set_data_numeric(x)
}
\arguments{
\item{x}{\code{data.frame}, raw data table with columns \code{sector} and years.}
}
\value{
\code{data.frame} of same form, but with \code{numeric} values in year columns
}
\description{
Tables containing raw data and NO, NO, NE, NA values are transfomred to numeric
with meta values carried in an table as attribute "data_meta"
}
