% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_variables}
\alias{get_variables}
\title{Extract ccmgu information based on variableId}
\usage{
get_variables(rms, ids)
}
\arguments{
\item{rms}{list, \code{remis} object}

\item{ids}{integer, variableId}
}
\value{
data.frame containing integer variableId and text description of each
corresponding ccmug id
}
\description{
Extract ccmgu information based on variableId
}
