% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_variables}
\alias{get_variables}
\title{Extract ccmgu information based on variableId}
\usage{
get_variables(rms, variable_id)
}
\arguments{
\item{rms}{list, \code{remis} object}

\item{variable_id}{numeric, \code{annexOne} or \code{nonAnnexOne} \code{variableId}s from \code{\link[=rem_init]{rem_init()}}\verb{$variables}}
}
\value{
data.frame containing variableId and text description of each
corresponding \emph{ccmug} id
}
\description{
Extract ccmgu information based on variableId
}
\examples{
\donttest{
rem <- rem_init()

rem$variables$annexOne$variableId[25]

get_variables(
  rem,
  rem$variables$annexOne$variableId[25])


}
}
