% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{parse_raw}
\alias{parse_raw}
\title{Readable query results}
\usage{
parse_raw(raw, rms)
}
\arguments{
\item{raw}{list, response from POST request}

\item{rms}{list, \code{remis} object}
}
\value{
data.frame
}
\description{
This function combines numeric/character responses from queries with their
respective id values from parties, years and variables.
}
