% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{post_parse}
\alias{post_parse}
\title{POST and parse objects}
\usage{
post_parse(x, url = "api/records/flexible-queries/", query)
}
\arguments{
\item{x}{list, remis object}

\item{url}{character, api path for flex query}
}
\value{
data.frame, parsed from json
}
\description{
POST and parse objects
}
