% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{rem_init}
\alias{rem_init}
\title{Initialize remis}
\usage{
rem_init(base_url = meta$api_base_url)
}
\arguments{
\item{base_url}{character, base url for queries}
}
\value{
list
}
\description{
Get meta data for the Flex Query API and initialize the API.
}
\examples{
\donttest{
rem <- rem_init()
names(rem)


}
}
